+++ app/src/main/java/com/hhp227/knu_minigroup/fragment/DCDormMealFragment.java
@@ -1,35 +1,22 @@
 package com.hhp227.knu_minigroup.fragment;
 
 import android.os.Bundle;
-import android.text.Html;
-import android.text.SpannableString;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+import android.widget.Toast;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import com.android.volley.Request;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.VolleyLog;
-import com.android.volley.toolbox.StringRequest;
-import com.hhp227.knu_minigroup.app.AppController;
-import com.hhp227.knu_minigroup.app.EndPoint;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
+
 import com.hhp227.knu_minigroup.databinding.FragmentDormmealBinding;
-import org.xmlpull.v1.XmlPullParser;
-import org.xmlpull.v1.XmlPullParserException;
-import org.xmlpull.v1.XmlPullParserFactory;
+import com.hhp227.knu_minigroup.viewmodel.DCDormMealViewModel;
 
-import java.io.IOException;
-import java.io.StringReader;
-import java.util.ArrayList;
-
-// TODO
 public class DCDormMealFragment extends Fragment {
-    private static final String TAG = "문화관 식단표";
-
     private TextView[] mMenuView;
 
     private FragmentDormmealBinding mBinding;
@@ -39,12 +26,7 @@ public class DCDormMealFragment extends Fragment {
     }
 
     @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         mBinding = FragmentDormmealBinding.inflate(inflater, container, false);
         return mBinding.getRoot();
     }
@@ -52,51 +34,43 @@ public class DCDormMealFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
+        DCDormMealViewModel viewModel = new ViewModelProvider(this).get(DCDormMealViewModel.class);
         mMenuView = new TextView[] {
                 mBinding.breakfast,
                 mBinding.lunch,
                 mBinding.dinner
         };
-        String endPoint = EndPoint.URL_KNU_DORM_MEAL.replace("{ID}", "2");
 
-        AppController.getInstance().addToRequestQueue(new StringRequest(Request.Method.GET, endPoint, new Response.Listener<String>() {
+        viewModel.mState.observe(getViewLifecycleOwner(), new Observer<DCDormMealViewModel.State>() {
             @Override
-            public void onResponse(String response) {
-                ArrayList<String> arrayList = new ArrayList<>();
-
-                try {
-                    XmlPullParserFactory factory = XmlPullParserFactory.newInstance();
-                    XmlPullParser parser = factory.newPullParser();
-
-                    parser.setInput(new StringReader(response));
-                    int eventType = parser.getEventType();
-                    while (eventType != XmlPullParser.END_DOCUMENT) {
-                        switch (eventType) {
-                            case XmlPullParser.START_TAG:
-                                String startTag = parser.getName();
-                                if (startTag.equals("data"))
-                                    arrayList.add(new SpannableString(Html.fromHtml(parser.nextText())).toString());
-                                break;
-                        }
-                        eventType = parser.next();
-                    }
-                } catch (XmlPullParserException | IOException e) {
-                    e.printStackTrace();
+            public void onChanged(DCDormMealViewModel.State state) {
+                if (state.isLoading) {
+                    showProgressBar();
+                } else if (!state.list.isEmpty()) {
+                    hideProgressBar();
+                    for (int i = 0; i < mMenuView.length; i++)
+                        mMenuView[i].setText(state.list.get(i));
+                } else if (state.message != null && !state.message.isEmpty()) {
+                    hideProgressBar();
+                    Toast.makeText(requireContext(), state.message, Toast.LENGTH_LONG).show();
                 }
-                for (int i = 0; i < mMenuView.length; i++)
-                    mMenuView[i].setText(arrayList.size() > 0 ? arrayList.get(i) : "등록된 식단이 없습니다.");
             }
-        }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-                VolleyLog.e(TAG, error.getMessage());
-            }
-        }));
+        });
     }
 
     @Override
     public void onDestroyView() {
         super.onDestroyView();
         mBinding = null;
+    }
+
+    private void showProgressBar() {
+        if (mBinding.progressBar.getVisibility() == View.GONE)
+            mBinding.progressBar.setVisibility(View.VISIBLE);
+    }
+
+    private void hideProgressBar() {
+        if (mBinding.progressBar.getVisibility() == View.VISIBLE)
+            mBinding.progressBar.setVisibility(View.GONE);
     }
 }
