+++ app/src/main/java/com/hhp227/knu_minigroup/fragment/DaeguSeatFragment.java
@@ -1,59 +1,34 @@
 package com.hhp227.knu_minigroup.fragment;
 
-import android.app.ProgressDialog;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
-import com.android.volley.Request;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.VolleyLog;
-import com.android.volley.toolbox.JsonObjectRequest;
+
+import com.google.android.material.snackbar.Snackbar;
 import com.hhp227.knu_minigroup.adapter.SeatListAdapter;
-import com.hhp227.knu_minigroup.app.AppController;
-import com.hhp227.knu_minigroup.app.EndPoint;
 import com.hhp227.knu_minigroup.databinding.FragmentSeatBinding;
-import com.hhp227.knu_minigroup.dto.SeatItem;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
+import com.hhp227.knu_minigroup.viewmodel.DaeguSeatViewModel;
 
-import java.util.ArrayList;
-import java.util.List;
-
-// TODO
 public class DaeguSeatFragment extends Fragment {
-    private static final String TAG = "대구 열람실좌석";
-
-    private boolean mIsRefresh;
-
-    private List<SeatItem> mSeatItemList;
-
-    private ProgressDialog mProgressDialog;
-
     private SeatListAdapter mAdapter;
 
     private FragmentSeatBinding mBinding;
 
-    public DaeguSeatFragment() {
-    }
+    private DaeguSeatViewModel mViewModel;
 
     public static DaeguSeatFragment newInstance() {
         return new DaeguSeatFragment();
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
     }
 
     @Override
@@ -65,9 +40,8 @@ public class DaeguSeatFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        mSeatItemList = new ArrayList<>();
-        mAdapter = new SeatListAdapter(mSeatItemList);
-        mProgressDialog = new ProgressDialog(getActivity());
+        mViewModel = new ViewModelProvider(this).get(DaeguSeatViewModel.class);
+        mAdapter = new SeatListAdapter(mViewModel.mSeatItemList);
 
         mBinding.recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
         mBinding.recyclerView.setAdapter(mAdapter);
@@ -77,17 +51,26 @@ public class DaeguSeatFragment extends Fragment {
                 new Handler(Looper.getMainLooper()).postDelayed(new Runnable() {
                     @Override
                     public void run() {
-                        mIsRefresh = true;
-                        fetchDataTask();
+                        mViewModel.refresh();
                         mBinding.srl.setRefreshing(false);
                     }
                 }, 1000);
             }
         });
-        mProgressDialog.setCancelable(false);
-        mProgressDialog.setMessage("불러오는중...");
-        showProgressDialog();
-        fetchDataTask();
+        mViewModel.mState.observe(getViewLifecycleOwner(), new Observer<DaeguSeatViewModel.State>() {
+            @Override
+            public void onChanged(DaeguSeatViewModel.State state) {
+                if (state.isLoading) {
+                    showProgressBar();
+                } else if (state.isSuccess) {
+                    hideProgressBar();
+                    mAdapter.notifyDataSetChanged();
+                } else if (!state.message.isEmpty()) {
+                    hideProgressBar();
+                    Snackbar.make(requireView(), state.message, Snackbar.LENGTH_LONG).show();
+                }
+            }
+        });
     }
 
     @Override
@@ -96,63 +79,13 @@ public class DaeguSeatFragment extends Fragment {
         mBinding = null;
     }
 
-    private void fetchDataTask() {
-        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET, EndPoint.URL_KNULIBRARY_SEAT.replace("{ID}", "1"), null, new Response.Listener<JSONObject>() {
-            @Override
-            public void onResponse(JSONObject response) {
-                try {
-                    JSONObject jsonObject = response.getJSONObject("data");
-                    JSONArray jsonArray = jsonObject.getJSONArray("list");
-
-                    for (int i = 0; i < jsonArray.length(); i++) {
-                        JSONObject data = jsonArray.getJSONObject(i);
-                        int id = data.getInt("id");
-                        String name = data.getString("name");
-                        int total = data.getInt("activeTotal");
-                        int occupied = data.getInt("occupied");
-                        int available = data.getInt("available");
-                        String[] disable = null;
-
-                        try {
-                            JSONObject disablePeriod = data.getJSONObject("disablePeriod");
-                            disable = new String[3];
-                            disable[0] = disablePeriod.getString("name");
-                            disable[1] = disablePeriod.getString("beginTime");
-                            disable[2] = disablePeriod.getString("endTime");
-                        } catch (JSONException e) {
-                            Log.e(TAG, e.getMessage());
-                        }
-
-                        SeatItem listItem = new SeatItem(id, name, total, occupied, available, disable);
-
-                        if (!mIsRefresh)
-                            mSeatItemList.add(listItem);
-                        else
-                            mSeatItemList.set(i, listItem);
-                    }
-                    mAdapter.notifyDataSetChanged();
-                } catch (JSONException e) {
-                    Log.e(TAG, e.getMessage());
-                }
-                hideProgressDialog();
-            }
-        }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-                VolleyLog.e(TAG, error.getMessage());
-                hideProgressDialog();
-            }
-        });
-        AppController.getInstance().addToRequestQueue(jsonObjectRequest);
+    private void showProgressBar() {
+        if (mBinding.progressCircular.getVisibility() == View.GONE)
+            mBinding.progressCircular.setVisibility(View.VISIBLE);
     }
 
-    private void showProgressDialog() {
-        if (!mProgressDialog.isShowing())
-            mProgressDialog.show();
-    }
-
-    private void hideProgressDialog() {
-        if (mProgressDialog.isShowing())
-            mProgressDialog.dismiss();
+    private void hideProgressBar() {
+        if (mBinding.progressCircular.getVisibility() == View.VISIBLE)
+            mBinding.progressCircular.setVisibility(View.GONE);
     }
 }
