+++ app/src/main/java/com/hhp227/knu_minigroup/fragment/SangjuSeatFragment.java
@@ -1,60 +1,33 @@
 package com.hhp227.knu_minigroup.fragment;
 
-import android.app.ProgressDialog;
 import android.os.Bundle;
 import android.os.Handler;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
-import com.android.volley.Request;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.VolleyLog;
-import com.android.volley.toolbox.JsonObjectRequest;
-import com.hhp227.knu_minigroup.R;
+
+import com.google.android.material.snackbar.Snackbar;
 import com.hhp227.knu_minigroup.adapter.SeatListAdapter;
-import com.hhp227.knu_minigroup.app.AppController;
-import com.hhp227.knu_minigroup.app.EndPoint;
 import com.hhp227.knu_minigroup.databinding.FragmentSeatBinding;
-import com.hhp227.knu_minigroup.dto.SeatItem;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
+import com.hhp227.knu_minigroup.viewmodel.SangjuSeatViewModel;
 
-import java.util.ArrayList;
-import java.util.List;
-
-// TODO
 public class SangjuSeatFragment extends Fragment {
-    private static final String TAG = "상주 열람실좌석";
-
-    private boolean mIsRefresh;
-
-    private List<SeatItem> mSeatItemList;
-
-    private ProgressDialog mProgressDialog;
-
     private SeatListAdapter mAdapter;
 
     private FragmentSeatBinding mBinding;
 
-    public SangjuSeatFragment() {
-    }
+    private SangjuSeatViewModel mViewModel;
 
     public static SangjuSeatFragment newInstance() {
         return new SangjuSeatFragment();
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
     }
 
     @Override
@@ -66,9 +39,8 @@ public class SangjuSeatFragment extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        mSeatItemList = new ArrayList<>();
-        mAdapter = new SeatListAdapter(mSeatItemList);
-        mProgressDialog = new ProgressDialog(getActivity());
+        mViewModel = new ViewModelProvider(this).get(SangjuSeatViewModel.class);
+        mAdapter = new SeatListAdapter(mViewModel.mSeatItemList);
 
         mBinding.recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
         mBinding.recyclerView.setAdapter(mAdapter);
@@ -78,17 +50,26 @@ public class SangjuSeatFragment extends Fragment {
                 new Handler().postDelayed(new Runnable() {
                     @Override
                     public void run() {
-                        mIsRefresh = true;
-                        fetchData();
+                        mViewModel.refresh();
                         mBinding.srl.setRefreshing(false);
                     }
                 }, 1000);
             }
         });
-        mProgressDialog.setCancelable(false);
-        mProgressDialog.setMessage("불러오는중...");
-        showProgressDialog();
-        fetchData();
+        mViewModel.mState.observe(getViewLifecycleOwner(), new Observer<SangjuSeatViewModel.State>() {
+            @Override
+            public void onChanged(SangjuSeatViewModel.State state) {
+                if (state.isLoading) {
+                    showProgressBar();
+                } else if (state.isSuccess) {
+                    hideProgressBar();
+                    mAdapter.notifyDataSetChanged();
+                } else if (!state.message.isEmpty()) {
+                    hideProgressBar();
+                    Snackbar.make(requireView(), state.message, Snackbar.LENGTH_LONG).show();
+                }
+            }
+        });
     }
 
     @Override
@@ -97,63 +78,13 @@ public class SangjuSeatFragment extends Fragment {
         mBinding = null;
     }
 
-    private void fetchData() {
-        AppController.getInstance().addToRequestQueue(new JsonObjectRequest(Request.Method.GET, EndPoint.URL_KNULIBRARY_SEAT.replace("{ID}", "2"), null, new Response.Listener<JSONObject>() {
-            @Override
-            public void onResponse(JSONObject response) {
-                try {
-                    JSONObject jsonObject = response.getJSONObject("data");
-                    JSONArray jsonArray = jsonObject.getJSONArray("list");
-
-                    for (int i = 0; i < jsonArray.length(); i++) {
-                        JSONObject data = jsonArray.getJSONObject(i);
-                        String[] disable = null;
-
-                        try {
-                            JSONObject disablePeriod = data.getJSONObject("disablePeriod");
-                            disable = new String[3];
-                            disable[0] = disablePeriod.getString("name");
-                            disable[1] = disablePeriod.getString("beginTime");
-                            disable[2] = disablePeriod.getString("endTime");
-                        } catch (JSONException e) {
-                            Log.e(TAG, e.getMessage());
-                        }
-
-                        SeatItem listItem = new SeatItem(
-                                data.getInt("id"),
-                                data.getString("name"),
-                                data.getInt("activeTotal"),
-                                data.getInt("occupied"),
-                                data.getInt("available"),
-                                disable);
-
-                        if (!mIsRefresh)
-                            mSeatItemList.add(listItem);
-                        else
-                            mSeatItemList.set(i, listItem);
-                    }
-                    mAdapter.notifyDataSetChanged();
-                } catch (JSONException e) {
-                    Log.e(TAG, e.getMessage());
-                }
-                hideProgressDialog();
-            }
-        }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-                VolleyLog.e(TAG, error.getMessage());
-                hideProgressDialog();
-            }
-        }));
+    private void showProgressBar() {
+        if (mBinding.progressCircular.getVisibility() == View.GONE)
+            mBinding.progressCircular.setVisibility(View.VISIBLE);
     }
 
-    private void showProgressDialog() {
-        if (!mProgressDialog.isShowing())
-            mProgressDialog.show();
-    }
-
-    private void hideProgressDialog() {
-        if (mProgressDialog.isShowing())
-            mProgressDialog.dismiss();
+    private void hideProgressBar() {
+        if (mBinding.progressCircular.getVisibility() == View.VISIBLE)
+            mBinding.progressCircular.setVisibility(View.GONE);
     }
 }
