+++ app/src/main/res/layout/fragment_dormmeal.xml
@@ -1,13 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_height="match_parent"
+    android:layout_width="match_parent">
+
+    <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#eeeeee"
         android:orientation="vertical">
 
-    <androidx.cardview.widget.CardView
+        <androidx.cardview.widget.CardView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_weight="1"
@@ -18,12 +22,12 @@
             android:layout_marginBottom="4dp"
             app:cardCornerRadius="4dp">
 
-        <RelativeLayout
+            <RelativeLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:padding="16dp">
 
-            <TextView
+                <TextView
                     android:id="@+id/breakfast_title"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -35,7 +39,7 @@
                     android:textAppearance="?android:attr/textAppearanceMedium"
                     android:textSize="20dp"/>
 
-            <LinearLayout
+                <LinearLayout
                     android:id="@+id/breakfast_below_line"
                     android:layout_width="match_parent"
                     android:layout_height="2px"
@@ -43,25 +47,25 @@
                     android:background="@android:color/darker_gray"
                     android:orientation="vertical"/>
 
-            <ScrollView
+                <ScrollView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_alignStart="@+id/breakfast_below_line"
                     android:layout_below="@+id/breakfast_below_line">
 
-                <TextView
+                    <TextView
                         android:id="@+id/breakfast"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:textSize="16dp"/>
 
-            </ScrollView>
+                </ScrollView>
 
-        </RelativeLayout>
+            </RelativeLayout>
 
-    </androidx.cardview.widget.CardView>
+        </androidx.cardview.widget.CardView>
 
-    <androidx.cardview.widget.CardView
+        <androidx.cardview.widget.CardView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_weight="1"
@@ -72,12 +76,12 @@
             android:layout_marginBottom="4dp"
             app:cardCornerRadius="4dp">
 
-        <RelativeLayout
+            <RelativeLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:padding="16dp">
 
-            <TextView
+                <TextView
                     android:id="@+id/lunch_title"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -89,7 +93,7 @@
                     android:textAppearance="?android:attr/textAppearanceMedium"
                     android:textSize="20dp"/>
 
-            <LinearLayout
+                <LinearLayout
                     android:id="@+id/lunch_below_line"
                     android:layout_width="match_parent"
                     android:layout_height="2px"
@@ -97,25 +101,25 @@
                     android:background="@android:color/darker_gray"
                     android:orientation="vertical"/>
 
-            <ScrollView
+                <ScrollView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_alignStart="@+id/lunch_below_line"
                     android:layout_below="@+id/lunch_below_line">
 
-                <TextView
+                    <TextView
                         android:id="@+id/lunch"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:textSize="16dp"/>
 
-            </ScrollView>
+                </ScrollView>
 
-        </RelativeLayout>
+            </RelativeLayout>
 
-    </androidx.cardview.widget.CardView>
+        </androidx.cardview.widget.CardView>
 
-    <androidx.cardview.widget.CardView
+        <androidx.cardview.widget.CardView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
@@ -126,12 +130,12 @@
             android:layout_marginBottom="4dp"
             app:cardCornerRadius="4dp">
 
-        <RelativeLayout
+            <RelativeLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:padding="16dp">
 
-            <TextView
+                <TextView
                     android:id="@+id/dinner_title"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -143,7 +147,7 @@
                     android:textAppearance="?android:attr/textAppearanceMedium"
                     android:textSize="20dp"/>
 
-            <LinearLayout
+                <LinearLayout
                     android:id="@+id/dinner_below_line"
                     android:layout_width="match_parent"
                     android:layout_height="2px"
@@ -151,22 +155,34 @@
                     android:background="@android:color/darker_gray"
                     android:orientation="vertical" />
 
-            <ScrollView
+                <ScrollView
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_alignStart="@+id/dinner_below_line"
                     android:layout_below="@+id/dinner_below_line">
 
-                <TextView
+                    <TextView
                         android:id="@+id/dinner"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
                         android:textSize="16dp"/>
 
-            </ScrollView>
+                </ScrollView>
 
-        </RelativeLayout>
+            </RelativeLayout>
 
-    </androidx.cardview.widget.CardView>
+        </androidx.cardview.widget.CardView>
 
-</LinearLayout>
+    </LinearLayout>
+
+    <ProgressBar
+        android:id="@+id/progress_bar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:visibility="gone"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"/>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
